\name{energy_emissions}
\alias{energy_emissions}
\title{Energy Emissions}
\description{
Converts electricity consumption into metric tons of CO2 emissions. Electricity emissions factors are sourced from the U.S. Energy Information Administration (2022).
}
\usage{
energy_emissions(consumption = 0, units = "kilowatt-hours")
}
\arguments{
  \item{consumption}{Amount of electricity consumed. Default is 0. The input may be a numeric value, a string value (containing only numbers), or a vector of numeric values, string values (containing only numbers), or both.}
  \item{units}{The units of electricity consumption data. Default units are kilowatt-hours, but gigawatt-hours and watt-hours can also be converted into CO2 emissions. The input must be a string value (either "gigawatt-hours", "kilowatt-hours", or "watt-hours"). Capitalization does not matter.}
}
\value{
A numeric value (or a vector of numeric values) representing the CO2 emissions (in metric tons) generated by the input electricity consumption.
}
\examples{
energy_emissions(2000)
energy_emissions(2000, "gigawatt-hours")
energy_emissions("2000")
energy_emissions(c(1000.33, 2000))
energy_emissions(consumption = 2000, units = "WATT-HOURS")
energy_emissions(consumption = c(1000.33, "2000"))
energy_emissions(data_frame$energy_consumption)
}