\name{fuel_emissions}
\alias{fuel_emissions}
\title{Fuel Emissions}
\description{
Converts fuel consumption into metric tons of CO2 emissions. Convertible fuels include gasoline and diesel. Fuel emissions factors are gathered from the U.S. Environmental Protection Agency (2010).
}
\usage{
fuel_emissions(gasoline = 0, diesel = 0, units = "gallons")
}
\arguments{
  \item{gasoline}{Amount of gasoline consumed. Default is 0. The input may be a numeric value, a string value (containing only numbers), or a vector of numeric values, string values (containing only numbers), or both.}
  \item{diesel}{Amount of diesel consumed. Default is 0. The input may be a numeric value, a string value (containing only numbers), or a vector of numeric values, string values (containing only numbers), or both.}
  \item{units}{Units of fuel consumption data (both gasoline and diesel). Default units are gallons, but liters and quarts can also be converted into CO2 emissions. The input must be a string (either "gallons", "liters", or "quarts"). Capitalization does not matter.}
}
\value{
A numeric value (or a vector of numeric values) representing the CO2 emissions (in metric tons) generated by the input fuel consumption.
}
\examples{
fuel_emissions(20, 300)
fuel_emissions("20", "300", "liters")
fuel_emissions(gasoline = 820, diesel = 22.25, units = "quarts")
fuel_emissions(diesel = 3.005)
fuel_emissions(gasoline = c(125, 1540), diesel = c(80, 2000))
fuel_emissions(data_frame$gasoline, data_frame$diesel)
}